+++ Simulation/AbstractAndInterfaces/RouteTime.java
@@ -11,5 +11,5 @@ import Simulation.Nouns.Truck;
 import java.util.PriorityQueue;
 
 public interface RouteTime {
-    public int route(PriorityQueue<Address> addresses, Truck truck);
+    public double route(PriorityQueue<Address> addresses, Truck truck);
 }
